global mtu settings
Former-commit-id: 64533fa8fa5070e7064da39096703aae39c6e2c2

+++ internal/client/wireguardControl/interfaceState/init.go
@@ -10,7 +10,7 @@ type interfaceState struct {
 
 	address string
 	network string
-	mtu     int
+	mtu     uint16
 	dns     []string
 	table   string
